+++ src/main/java/io/github/imfangs/dify/client/callback/ChatflowStreamCallback.java
@@ -32,6 +32,14 @@ public interface ChatflowStreamCallback extends ChatStreamCallback {
     }
 
     /**
+     * 节点重试事件
+     *
+     * @param event 事件数据
+     */
+    default void onNodeRetry(NodeRetryEvent event) {
+    }
+
+    /**
      * 工作流完成事件
      *
      * @param event 事件数据